+++ src/app/admin/components/admin-products/admin-products.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
-import { Location } from '@angular/common';
+import { Router } from '@angular/router';
 
 import { IProductModel } from '../../../product/models/iproduct-model';
 import { ProductsService } from '../../../product/services/products.service';
@@ -11,20 +10,15 @@ import { ProductsService } from '../../../product/services/products.service';
   styleUrls: ['./admin-products.component.less']
 })
 export class AdminProductsComponent implements OnInit {
-  products$: Observable<IProductModel[]>;
+  products: IProductModel[];
 
-  constructor(private location: Location, private productsService: ProductsService) { }
+  constructor(private router: Router, private productsService: ProductsService) { }
 
   ngOnInit(): void {
-    this.products$ = this.productsService.getProducts();
+    this.productsService.products$.subscribe(data => this.products = data);
   }
-
-  updateProductData(data: IProductModel): void {
-
-  }
-
 
   goBack(): void{
-    this.location.back();
+    this.router.navigateByUrl('/admin');
   }
 }