+++ src/app/product/product-list-component/product-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
+import { ActivatedRoute } from '@angular/router';
 
 import { ProductsService } from '../services/products.service';
 import { CartService } from '../../cart/services/cart.service';
@@ -11,43 +11,56 @@ import { ICartProductModel } from '../../cart/models/icart-product-model';
   templateUrl: './product-list.component.html'
 })
 export class ProductListComponent implements OnInit {
-  products$: Observable<IProductModel[]>;
+  products: IProductModel[];
+  private initedCartData = false;
 
-  constructor(private productsService: ProductsService, private cartService: CartService) { }
+  constructor( private productsService: ProductsService, private cartService: CartService, private route: ActivatedRoute ) { }
 
   ngOnInit(): void {
-    this.products$ = this.productsService.getProducts();
-    this.cartService.data$.subscribe(res => this.updateProductData(res.removedProducts.length ? res.removedProducts : res.cartProducts));
+    this.productsService.products$.subscribe(data => {
+      this.products = data;
+      if (this.checkCartProducts()) {
+        this.initedCartData = true;
+        this.updateProductData(this.cartService.getProductsArray());
+      }
+    });
+    
+    this.cartService.data$.subscribe(res => {
+      if (this.checkCartProducts()) {
+        this.initedCartData = true;
+      }
+      this.updateProductData(res.removedProducts.length ? res.removedProducts : res.cartProducts);
+    });
   }
 
-  addToCart(data: IProductModel): void {
+  private checkCartProducts(): boolean {
+    return this.products.length && !this.initedCartData;
+  }
+
+  addToCart(data: IProductModel, cartAddedQty: number): void {
     this.cartService.addProduct({
       id: data.id,
       name: data.name,
       imgSrc: data.imgSrc,
       price: data.price,
       quantity: data.cartAddedQty,
-      stockQty: data.stockQty
+      stockQty: data.stockQty,
     });
+
+    this.productsService.updateProduct(data);
   }
 
   updateProductData(data: ICartProductModel[]): void {
-    this.products$.subscribe(products => {
-      const cartProducts = products.map((product: IProductModel) => {
-        const productObj = {...product};
+    this.products.map((product: IProductModel) => {
+      const productObj = {...product};
 
-        data.forEach((item): void => {
-          if (productObj.id === item.id) {
-            productObj.stockQty = item.stockQty;
-          }
-        });
-        return productObj;
+      data.forEach((item): void => {
+        if (productObj.id === item.id) {
+          productObj.stockQty = item.stockQty;
+        }
       });
 
-      this.products$ = new Observable((observer) => {
-          observer.next(cartProducts);
-          observer.complete();
-      });
+      this.productsService.updateProduct(productObj);
     });
   }
 }