+++ src/app/product/services/products.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-
-import { Observable } from 'rxjs';
+import { BehaviorSubject } from 'rxjs';
 
 import { IProductModel } from '../models/iproduct-model';
 
@@ -10,16 +9,44 @@ import { IProductModel } from '../models/iproduct-model';
   providedIn: 'root'
 })
 export class ProductsService {
-
+  private products: IProductModel[] = [];
   private dataJson = '/assets/data.json';
+  private dataProducts: any;
+  products$: any;
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {
+    this.dataProducts = new BehaviorSubject(this.products);
+    this.products$ = this.dataProducts.asObservable();
 
-  getProducts(): Observable<IProductModel[]> {
-    return this.http.get<IProductModel[]>(this.dataJson);
+    this.http.get<IProductModel[]>(this.dataJson)
+      .subscribe((data: IProductModel[]) => {
+        this.products = data;
+
+        this.dataProducts.next(data);
+      });
   }
 
-  updateProductListData(): void {
+  getProductById(id: string): IProductModel {
+    return this.products.find((item: IProductModel): IProductModel => {
+      if (item.id === id) {
+        return item;
+      }
+    });
+  }
 
+  addProduct(data: IProductModel): void {
+    this.products.unshift(data);
+    this.dataProducts.next(this.products);
+  }
+
+  updateProduct(data: IProductModel): void {
+    this.products = this.products.map((item: IProductModel): IProductModel => {
+      if (item.id === data.id) {
+        item = data;
+      }
+      return item;
+    });
+
+    this.dataProducts.next(this.products);
   }
 }